Move data formatting logic for sewer water to a util file
